Completely refactor exception handling
I'd incorrectly implemented exception handling, by treating the presence of
Error tags in the Query reponse as being a fatal error, but this is incorrect.
Exceptions should be raised when the response from the Netaxept API is an
Exception. The Error should be treated differently, with a method available to
retrieve the error.